Refactor - split item/recipe to files, shrink body of main
