+++ src/crafting.rs
@@ -1,16 +1,13 @@
 use crate::api;
-use crate::gw2efficiency;
 use crate::config;
+use crate::recipe::Recipe;
+use crate::item::Item;
 use crate::money::Money;
-
-use serde::{Deserialize, Serialize};
 
 use num_rational::Ratio;
 use num_traits::Zero;
 
 use std::collections::{BTreeMap, HashMap, HashSet};
-use std::convert::TryFrom;
-use std::cmp::Ordering;
 
 #[derive(Debug, Copy, Clone, Eq, Hash, Ord, PartialEq, PartialOrd)]
 pub enum Source {
@@ -30,7 +27,7 @@ pub struct EstimatedCraftingCost {
 pub fn calculate_estimated_min_crafting_cost(
     item_id: u32,
     recipes_map: &HashMap<u32, Recipe>,
-    items_map: &HashMap<u32, api::Item>,
+    items_map: &HashMap<u32, Item>,
     tp_prices_map: &HashMap<u32, api::Price>,
     opt: &config::CraftingOptions,
 ) -> Option<EstimatedCraftingCost> {
@@ -135,22 +132,14 @@ impl CraftedItems {
                 continue;
             };
             if let Some(id) = recipe.id.filter(|id| !unknown_recipes.contains(id)) {
-                match recipe.source {
-                    RecipeSource::Purchasable | RecipeSource::Achievement => {
-                        // If we have no known recipes, assume we know none
-                        if known_recipes
-                            .as_ref()
+                if !recipe.is_automatic() {
+                    // If we have no known recipes, assume we know none
+                    if known_recipes
+                        .as_ref()
                             .filter(|recipes| recipes.contains(&id))
                             .is_none()
-                        {
-                            unknown_recipes.insert(id);
-                        }
-                    }
-                    // These aren't included in the API; assume you know them
-                    RecipeSource::Automatic | RecipeSource::Discoverable => {
-                        // TODO: instead, check if account has a char with the required crafting level
-                        // Would require a key with the characters scope. Still wouldn't detect
-                        // discoverable recipes, but would detect access to them
+                    {
+                        unknown_recipes.insert(id);
                     }
                 }
             }
@@ -219,7 +208,7 @@ fn calculate_precise_min_crafting_cost(
     item_id: u32,
     item_count: u32,
     recipes_map: &HashMap<u32, Recipe>,
-    items_map: &HashMap<u32, api::Item>,
+    items_map: &HashMap<u32, Item>,
     tp_listings_map: &mut BTreeMap<u32, ItemListings>,
     context: &mut PreciseCraftingCostContext,
     opt: &config::CraftingOptions,
@@ -376,7 +365,7 @@ fn calculate_precise_min_crafting_cost(
 pub fn calculate_crafting_profit(
     item_id: u32,
     recipes_map: &HashMap<u32, Recipe>,
-    items_map: &HashMap<u32, api::Item>,
+    items_map: &HashMap<u32, Item>,
     tp_listings_map: &HashMap<u32, api::ItemListings>,
     mut purchased_ingredients: Option<&mut HashMap<(u32, Source), PurchasedIngredient>>,
     opt: &config::CraftingOptions,
@@ -690,115 +679,6 @@ impl From<api::ItemListings> for ItemListings {
                 .collect(),
             pending_buy_quantity: 0,
         }
-    }
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub enum RecipeSource {
-    Automatic,
-    Discoverable,
-    Purchasable,
-    Achievement,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct Recipe {
-    pub id: Option<u32>,
-    pub output_item_id: u32,
-    pub output_item_count: u32,
-    pub disciplines: Vec<config::Discipline>,
-    pub ingredients: Vec<api::RecipeIngredient>,
-    source: RecipeSource,
-}
-
-impl From<api::Recipe> for Recipe {
-    fn from(recipe: api::Recipe) -> Self {
-        let source = if recipe.is_purchased() {
-            RecipeSource::Purchasable
-        } else if recipe.is_automatic() {
-            RecipeSource::Automatic
-        } else {
-            RecipeSource::Discoverable
-        };
-        Recipe {
-            id: Some(recipe.id),
-            output_item_id: recipe.output_item_id,
-            output_item_count: recipe.output_item_count,
-            disciplines: recipe.disciplines,
-            ingredients: recipe.ingredients,
-            source,
-        }
-    }
-}
-
-impl TryFrom<gw2efficiency::Recipe> for Recipe {
-    type Error = String;
-
-    fn try_from(recipe: gw2efficiency::Recipe) -> Result<Self, Self::Error> {
-        let output_item_count = if let Some(count) = recipe.output_item_count {
-            count
-        } else {
-            return Err(format!(
-                "Ignoring '{}'. Failed to parse 'output_item_count' as integer.",
-                recipe.name
-            ));
-        };
-        // Any disciplines _except_ Achievement can be counted as known
-        // While some regular discipline precursor recipes must be learned, the
-        // outputs appear to be account bound anyway, so won't be on TP.
-        // There are some useful Scribe WvW BPs in the data, so ignoring all
-        // normal discipline recipes would catch those too.
-        let source = if recipe
-            .disciplines
-            .contains(&config::Discipline::Achievement)
-        {
-            RecipeSource::Achievement
-        } else {
-            RecipeSource::Automatic
-        };
-        Ok(Recipe {
-            id: None,
-            output_item_id: recipe.output_item_id,
-            output_item_count,
-            disciplines: recipe.disciplines,
-            ingredients: recipe.ingredients,
-            source,
-        })
-    }
-}
-
-impl Recipe {
-    // see https://wiki.guildwars2.com/wiki/Category:Time_gated_recipes
-    // for a list of time gated recipes
-    // I've left Charged Quartz Crystals off the list, since they can
-    // drop from containers.
-    pub fn is_timegated(&self) -> bool {
-        self.output_item_id == 46740         // Spool of Silk Weaving Thread
-            || self.output_item_id == 46742  // Lump of Mithrillium
-            || self.output_item_id == 46744  // Glob of Elder Spirit Residue
-            || self.output_item_id == 46745  // Spool of Thick Elonian Cord
-            || self.output_item_id == 66913  // Clay Pot
-            || self.output_item_id == 66917  // Plate of Meaty Plant Food
-            || self.output_item_id == 66923  // Plate of Piquant Plan Food
-            || self.output_item_id == 67015  // Heat Stone
-            || self.output_item_id == 67377  // Vial of Maize Balm
-            || self.output_item_id == 79726  // Dragon Hatchling Doll Eye
-            || self.output_item_id == 79763  // Gossamer Stuffing
-            || self.output_item_id == 79790  // Dragon Hatchling Doll Hide
-            || self.output_item_id == 79795  // Dragon Hatchling Doll Adornments
-            || self.output_item_id == 79817  // Dragon Hatchling Doll Frame
-            || self.output_item_id == 43772 // Charged Quartz Crystal
-    }
-
-    pub fn sorted_ingredients(&self) -> Vec<&api::RecipeIngredient> {
-        let mut ingredients: Vec<&api::RecipeIngredient> = self.ingredients.iter().collect();
-        ingredients.sort_unstable_by(|a, b| {
-            match b.count.cmp(&a.count) {
-                Ordering::Equal => b.item_id.cmp(&a.item_id),
-                v => v,
-            }
-        });
-        ingredients
     }
 }
