+++ src/lib.rs
@@ -4,3 +4,5 @@ pub mod crafting;
 pub mod gw2efficiency;
 pub mod request;
 pub mod money;
+pub mod recipe;
+pub mod item;
